Refactor: 회원 정보 요청시 인증 서버의 회원정보까지 함께 응답 #127
