+++ src/main/java/com/nhnacademy/marketgg/server/dto/AuthInfo.java
@@ -7,17 +7,8 @@ import lombok.NoArgsConstructor;
 @Getter
 public class AuthInfo {
 
-    private String uuid;
     private String email;
     private String name;
     private String phoneNumber;
-    private boolean flag;
-
-    public void setUuid(String uuid) {
-        if (!flag) {
-            this.uuid = uuid;
-            flag = true;
-        }
-    }
 
 }
