+++ src/main/java/com/nhnacademy/marketgg/server/dto/MemberInfo.java
@@ -1,7 +1,6 @@
 package com.nhnacademy.marketgg.server.dto;
 
 import com.nhnacademy.marketgg.server.entity.Cart;
-import com.nhnacademy.marketgg.server.entity.MemberGrade;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
@@ -13,7 +12,7 @@ public class MemberInfo {
 
     private final Long id;
     private final Cart cart;
-    private final MemberGrade memberGrade;
+    private final String memberGrade;
     private final Character gender;
     private final LocalDate birthDate;
     private final LocalDateTime ggpassUpdatedAt;
