+++ src/main/java/com/nhnacademy/marketgg/server/dto/response/MemberResponse.java
@@ -1,6 +1,7 @@
 package com.nhnacademy.marketgg.server.dto.response;
 
-import com.nhnacademy.marketgg.server.entity.MemberGrade;
+import com.nhnacademy.marketgg.server.dto.AuthInfo;
+import com.nhnacademy.marketgg.server.dto.MemberInfo;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import lombok.Builder;
@@ -12,9 +13,22 @@ import lombok.RequiredArgsConstructor;
 @Getter
 public class MemberResponse {
 
-    private final MemberGrade memberGrade;
+    private final String email;
+    private final String name;
+    private final String phoneNumber;
+    private final String memberGrade;
     private final Character gender;
     private final LocalDate birthDay;
     private final LocalDateTime ggpassUpdatedAt;
+
+    public MemberResponse(AuthInfo authInfo, MemberInfo memberInfo) {
+        this.email = authInfo.getEmail();
+        this.name = authInfo.getName();
+        this.phoneNumber = authInfo.getPhoneNumber();
+        this.memberGrade = memberInfo.getMemberGrade();
+        this.gender = memberInfo.getGender();
+        this.birthDay = memberInfo.getBirthDate();
+        this.ggpassUpdatedAt = memberInfo.getGgpassUpdatedAt();
+    }
 
 }
