+++ src/main/java/com/nhnacademy/marketgg/server/repository/member/MemberRepositoryImpl.java
@@ -21,7 +21,7 @@ public class MemberRepositoryImpl extends QuerydslRepositorySupport implements M
             .innerJoin(member.cart)
             .where(member.uuid.eq(uuid))
             .select(Projections.constructor(MemberInfo.class,
-                member.id, member.cart, member.memberGrade, member.gender, member.birthDate, member.ggpassUpdatedAt))
+                member.id, member.cart, member.memberGrade.grade, member.gender, member.birthDate, member.ggpassUpdatedAt))
             .fetchOne();
 
         return Optional.ofNullable(memberInfo);
