+++ src/main/resources/application.yml
@@ -1,92 +1,3 @@
-#gg:
-#  gateway:
-#    origin: http://127.0.0.1:6060
-#
-#server:
-#  port: 7080
-#
-#spring:
-#  application:
-#    name: marketgg-shop
-#
-#  profiles:
-#    group:
-#      "local": "common, testdb"
-#      "dev": "common, devdb"
-#      "prod": "common, proddb"
-#
-#---
-#
-#spring:
-#  config:
-#    activate:
-#      on-profile: "common"
-#    import: "optional:configserver:http://133.186.209.70:8888"
-#
-#
-#  application:
-#    name: marketgg-shop
-#
-#  servlet:
-#    multipart:
-#      file-size-threshold: 1MB
-#      max-file-size: 100MB
-#      max-request-size: 100MB
-#
-#  elasticsearch:
-#    uris: 133.186.153.181:9200
-#    connection-timeout: 30
-#
-#  main:
-#    allow-bean-definition-overriding: true
-#
-#  jpa:
-#    database: mysql
-#    database-platform: org.hibernate.dialect.MySQL5InnoDBDialect
-#
-#---
-#
-#spring:
-#  config:
-#    activate:
-#      on-profile: "testdb"
-#
-#  h2:
-#    console:
-#      enabled: true
-#
-#  datasource:
-#    hikari:
-#      jdbc-url: jdbc:h2:mem:~/h2/testdb;MODE=MYSQL
-#    driver-class-name: org.h2.Driver
-#    username: sa
-#
-#  jpa:
-#    hibernate:
-#      ddl-auto: update
-#    show-sql: true
-#    properties:
-#      hibernate.format_sql: true
-#
-#---
-#
-#spring:
-#  config:
-#    activate:
-#      on-profile: "devdb"
-#
-#  datasource:
-#    driver-class-name: com.mysql.cj.jdbc.Driver
-#    url: jdbc:mysql://133.186.211.156:3306/marketgg_shop
-#    username: on7
-#    password: ?nk9Wa#^Bg)on71!*,^2=d
-#
-#decorator:
-#  datasource:
-#    p6spy:
-#      enable-logging: false
-
-
 spring:
   application:
     name: marketgg-shop
